Add double-experimental semver resolving in OCI
This is totally wrong and isn't how anything is supposed to work.
But it seems to work fine with `helm template` and `helm install` at least.

`helm pull` instead thinks the semver range from the user is the version
of the chart it grabbed, so the file is named badly.